requirements.txt
fastapi
uvicorn[standard]
ollama
python-dotenv
pandas
jinja2 